% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pse.R
\name{AddMetadata}
\alias{AddMetadata}
\title{AddMURPMetadata}
\usage{
AddMetadata(object, df)
}
\arguments{
\item{object}{MGPfact object}

\item{df}{attribute data frame to be added}
}
\description{
add attribute dataframe to object
}
