% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{AdjustPvalue}
\alias{AdjustPvalue}
\title{AdjustPvalue}
\usage{
AdjustPvalue(cs, ct_method, adjust_method)
}
\arguments{
\item{cs}{MGPfact-gene object}

\item{ct_method}{dge method}

\item{adjust_method}{filter threshold

ct_method = "lm"
sig_method = "pvalue"}
}
\description{
adjust pvalue using different method
}
