% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{assignSettings}
\alias{assignSettings}
\title{assignSettings}
\usage{
assignSettings(object, slotName = NULL, values = NULL)
}
\arguments{
\item{object}{MGPfact object}

\item{slotName}{slot name}

\item{values}{slot value}
}
\description{
add a single parameter
}
