% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{assignSettingsMulti}
\alias{assignSettingsMulti}
\title{assignSettingsMulti}
\usage{
assignSettingsMulti(object, slot)
}
\arguments{
\item{object}{MGPfact object}

\item{slot}{a list containing parameters, must be named with the parameter name}
}
\description{
add multiple parameters
}
