% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{$<-,MGPfact-method}
\alias{$<-,MGPfact-method}
\title{$-MGPfact-method}
\usage{
\S4method{$}{MGPfact}(x, name) <- value
}
\arguments{
\item{x}{An MGPfact object}

\item{name}{The column name}

\item{value}{The column values}
}
\description{
Replace or add a column in the metadata
}
