% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{compute_pval}
\alias{compute_pval}
\title{compute_pval}
\usage{
compute_pval(dat, mc, attr_name, value_name, group_name, test_method, seq = 1)
}
\arguments{
\item{dat}{data used for calculating P-values, including group column.}

\item{mc}{A list of length-2 vectors. The entries in the vector are
either the names of 2 values on the x-axis or the 2 integers
that correspond to the index of the groups of interest, to be compared}

\item{attr_name}{column name for splitting the data frame}

\item{value_name}{Numeric column for plotting.}

\item{group_name}{column name for data grouping within each data frame}

\item{test_method}{statistical test method}

\item{seq}{the interval between p-values, used for plotting}
}
\description{
compte pvalue
}
