% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{ComputeAOV2EffectGene}
\alias{ComputeAOV2EffectGene}
\title{ComputeAOV2EffectGene}
\usage{
ComputeAOV2EffectGene(object, adjust_method = "bonferroni", save_mod = FALSE)
}
\arguments{
\item{object}{MGPfact object}

\item{adjust_method}{correction method, a ‘character’ string.  Can be abbreviated.}

\item{save_mod}{Logical value, whether save result in mgpfact object}
}
\description{
Utilize ANOVA to calculate gene differential
expression between branches after bifurcation within each trajectory.
}
