% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{ComputeCorGeneT}
\alias{ComputeCorGeneT}
\title{ComputeMeanDiff}
\usage{
ComputeCorGeneT(object, method = "spearman", adjust_method = "bonferroni")
}
\arguments{
\item{object}{MGPfact object}

\item{method}{method for calculating correlation coefficients}

\item{adjust_method}{correction method, a ‘character’ string.  Can be abbreviated.}
}
\description{
Calculate the correlation coefficient between gene expression and trajectory scores
}
