% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{ComputeGeneWeight}
\alias{ComputeGeneWeight}
\title{ComputeWeightGene}
\usage{
ComputeGeneWeight(object = NULL)
}
\arguments{
\item{object}{MGPfact object}
}
\description{
Calculate gene weights across different trajectories
}
