% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{ComputeInteractionEffectGene}
\alias{ComputeInteractionEffectGene}
\title{ComputeInteractionEffectGene}
\usage{
ComputeInteractionEffectGene(
  object,
  adjust_method = "bonferroni",
  save_mod = FALSE
)
}
\arguments{
\item{object}{MGPfact object}

\item{adjust_method}{correction method, a ‘character’ string.  Can be abbreviated.}

\item{save_mod}{Logical value, whether save result in mgpfact object}
}
\description{
Use lm (linear model) to calculate gene differential expression between
different branches within each trajectory,
considering branch labels and pseudotime.
}
