% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ComputeLogpdf}
\alias{ComputeLogpdf}
\title{ComputeLogpdf}
\usage{
ComputeLogpdf(object, import_sim = FALSE, julia_home = NULL)
}
\arguments{
\item{object}{MGPfact object}

\item{import_sim}{logical value, whether import result from julia environment}

\item{julia_home}{Julia's bin path}
}
\description{
compute probability density values for different parameters
}
