% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{ComputeMeanDiff}
\alias{ComputeMeanDiff}
\title{ComputeMeanDiff}
\usage{
ComputeMeanDiff(object, method = "wilcox.test", adjust_method = "bonferroni")
}
\arguments{
\item{object}{MGPfact object}

\item{method}{Significance testing method}

\item{adjust_method}{correction method, a ‘character’ string.  Can be abbreviated.}
}
\description{
Calculate the mean differences between branches for different genes.
}
