% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{ComputeSigData}
\alias{ComputeSigData}
\title{ComputeSigData}
\usage{
ComputeSigData(ct, cs, sig_label, ct_method, sig_method, sig_th)
}
\arguments{
\item{ct}{MGPfact object}

\item{cs}{MGPfact-gene object}

\item{sig_label}{dge factor}

\item{ct_method}{dge method}

\item{sig_method}{statistics label}

\item{sig_th}{filter threshold

sig = TRUE
ct_method = "lm"
sig_method = "pvalue"
sig_th = 0.05}
}
\description{
get new sdata and cl according to differential gene
}
