% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{ComputeSigGene}
\alias{ComputeSigGene}
\title{ComputeSigGene}
\usage{
ComputeSigGene(cs, ct_method, sig_method, sig_th)
}
\arguments{
\item{cs}{MGPfact-gene object}

\item{ct_method}{dge method}

\item{sig_method}{statistics label}

\item{sig_th}{filter threshold

sig = TRUE
ct_method = "lm"
sig_method = "pvalue"
sig_th = 0.05}
}
\description{
compute significance genes with lm or aov
}
