% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pse.R
\name{CorViolinPlot}
\alias{CorViolinPlot}
\title{CorViolinPlot}
\usage{
CorViolinPlot(
  truetime = NULL,
  pseudotime = NULL,
  title = "TrueTime ~ PseudoTime",
  ftheme = NULL
)
}
\arguments{
\item{truetime}{sampling time points}

\item{pseudotime}{pseudotime}

\item{title}{plot title}

\item{ftheme}{ggplot theme}
}
\description{
draw the correlation between truetime and pseudotime in geom_violin
}
