% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateDimReducObject}
\alias{CreateDimReducObject}
\title{Create DimReducs Object}
\usage{
CreateDimReducObject(pca = NULL, tsne = NULL, umap = NULL, dm = NULL)
}
\arguments{
\item{pca}{pca result, from 'RUNPCA'}

\item{tsne}{tsne result, from 'RUNtSNE'}

\item{umap}{umap result, from 'RUNUMAP'}

\item{dm}{dm result, from 'RUNDM'}
}
\description{
create dimensionality reduction object
}
