% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateMGPfactObject}
\alias{CreateMGPfactObject}
\title{Create MGPfact Object}
\usage{
CreateMGPfactObject(
  data_matrix = NULL,
  count_matrix = NULL,
  datasetTitle = "project",
  dir = NULL,
  OptimResult = NULL,
  MURP = NULL,
  DimReducs = CreateDimReducObject(),
  BeautGene = list(),
  MetaData = NULL,
  Tree = NULL,
  GPR = NULL,
  Command = NULL
)
}
\arguments{
\item{data_matrix}{gene expression matrix, maybe the expression matrix after normalization or centralization}

\item{count_matrix}{original count matrix}

\item{datasetTitle}{project name}

\item{dir}{path to create folder}

\item{OptimResult}{optim result from julia package MGPfact}

\item{MURP}{MURP result from R package MURP}

\item{DimReducs}{dimensionality reduction results}

\item{BeautGene}{results related to genes}

\item{MetaData}{cell information}

\item{Tree}{trajectory toplog}

\item{GPR}{gaussian process regression}

\item{Command}{execution log}
}
\description{
create new object of MGPfact
}
