% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreatePseOptimResultObject}
\alias{CreatePseOptimResultObject}
\title{Create PseOptimResult Object}
\usage{
CreatePseOptimResultObject(
  sample_value = NULL,
  dimnames = NULL,
  burnin = NULL,
  chains = NULL,
  iter = NULL,
  hasinits = FALSE,
  inits = list(),
  iter_chain_cor = list(),
  sortindex_cor = list(),
  t_pred = list(),
  diagnostic = list(),
  logpdf = list()
)
}
\arguments{
\item{sample_value}{sampling results for each iteration}

\item{dimnames}{names of sampling value}

\item{burnin}{burning period}

\item{chains}{chain number,vector}

\item{iter}{iteration number}

\item{hasinits}{is there an initial value}

\item{inits}{inital value}

\item{iter_chain_cor}{Correlation of pseudo-time results between different Markov chains}

\item{sortindex_cor}{(ready to delete)}

\item{t_pred}{pseudotime list}

\item{diagnostic}{convergence judgment results}

\item{logpdf}{The objective function value of each optimization}
}
\description{
Create PseOptimResult Object
}
