% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateSettings}
\alias{CreateSettings}
\title{Create DimReducs Object}
\usage{
CreateSettings(datasetTitle = NULL, settings = NULL)
}
\arguments{
\item{datasetTitle}{project name}

\item{settings}{parameter settings}
}
\description{
create settings object
}
