% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateTrackOptimResultObject}
\alias{CreateTrackOptimResultObject}
\title{Create TrackOptimResult Object}
\usage{
CreateTrackOptimResultObject(
  sample_value = NULL,
  dimnames = NULL,
  burnin = NULL,
  chains = NULL,
  iter = NULL,
  hasinits = FALSE,
  inits = list(),
  diagnostic = list(),
  logpdf = list(),
  gene_weight = list()
)
}
\arguments{
\item{sample_value}{sampling results for each iteration}

\item{dimnames}{names of sampling value}

\item{burnin}{burning period}

\item{chains}{chain number,vector}

\item{iter}{iteration number}

\item{hasinits}{is there an initial value}

\item{inits}{inital value}

\item{diagnostic}{convergence judgment results}

\item{logpdf}{The objective function value of each optimization}

\item{gene_weight}{gene weight}
}
\description{
create track julia result object
}
