% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{dim,MGPfact-method}
\alias{dim,MGPfact-method}
\title{dim}
\usage{
\S4method{dim}{MGPfact}(x)
}
\arguments{
\item{x}{An MGPfact object.}
}
\description{
Returns the expression matrix in the MGPfact object and the dimensions of the downsampling matrix
}
