% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{DimReduc-class}
\alias{DimReduc-class}
\alias{DimReducs}
\title{The DimReduc-class}
\description{
The DimReduc-class
}
\section{Slots}{

\describe{
\item{\code{PCA}}{pca result}

\item{\code{tSNE}}{tsne result}

\item{\code{DM}}{diffusion map result}

\item{\code{UMAP}}{umap result}
}}

