% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ExceptSet}
\alias{ExceptSet}
\title{ExceptSet}
\usage{
ExceptSet(set_list = NULL)
}
\arguments{
\item{set_list}{List containing multiple sets.}
}
\description{
For multiple sets within a list, list the elements that differ between any two sets.
}
