% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gaussian_process_regression}
\alias{gaussian_process_regression}
\title{gaussian_process_regression}
\usage{
gaussian_process_regression(s, P, gene_expr, n.samples = 100)
}
\arguments{
\item{s}{covariance matrix}

\item{P}{murp number}

\item{gene_expr}{expression vector of any gene in the 'murp' dataset.}

\item{n.samples}{The number of points included in the regression results}
}
\description{
Use Gaussian process regression to calculate the smooth expression of any gene/trajectory score along a trajectory.
}
