% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{GeneBar}
\alias{GeneBar}
\title{GeneBar}
\usage{
GeneBar(
  object,
  gene = NULL,
  group = NULL,
  exprs = NULL,
  murp = NULL,
  expr_method = "mean",
  barcolor = colorRampPalette(pal_rickandmorty()(12))(12),
  errorbar = FALSE
)
}
\arguments{
\item{object}{MGPfact object}

\item{gene}{any gene}

\item{group}{group name}

\item{exprs}{expression matrix}

\item{murp}{murp result}

\item{expr_method}{mean/sum expression}

\item{barcolor}{color of bar}

\item{errorbar}{logical value, whether to add errorbar}
}
\description{
Bar plot of gene expression in different branches.
}
