% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{GeneBoxPhase}
\alias{GeneBoxPhase}
\title{GeneBoxPhase}
\usage{
GeneBoxPhase(
  object,
  gene = NULL,
  aspect = "cell",
  zscore = TRUE,
  test_method = "t.test",
  save = TRUE,
  label_y = NULL,
  length_width = 0.7,
  y_range = NULL
)
}
\arguments{
\item{object}{mgpfact object}

\item{gene}{gene name}

\item{aspect}{murp / cell}

\item{zscore}{logical value, whether to zscore}

\item{test_method}{t.test / wilcox.test}

\item{save}{logical value, whether to save pdf}

\item{label_y}{Annotate the starting position of the p-value on the y-axis}

\item{length_width}{Annotate the width of a single p-value}

\item{y_range}{The range of the y-axis}
}
\description{
violin plot of gene expression in different branches.
}
