% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{GeneCurvePlot}
\alias{GeneCurvePlot}
\title{GeneCurvePlot}
\usage{
GeneCurvePlot(
  object,
  gene,
  col = NULL,
  pointSize = 5,
  pointAlpha = 0.2,
  lineSize = 2
)
}
\arguments{
\item{object}{mgpfact object}

\item{gene}{gene name}

\item{col}{color column}
}
\description{
violin plot of gene expression in different branches.
}
