% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{GeneExprMURP}
\alias{GeneExprMURP}
\title{GeneExprMURP}
\usage{
GeneExprMURP(gene = NULL, expr = NULL, murp = NULL, methods = NULL)
}
\arguments{
\item{gene}{any gene}

\item{expr}{expression matrix}

\item{murp}{murp result}

\item{methods}{mean/sum expression}
}
\description{
Obtain the expression vector of a gene in MURP
}
