% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{GeneVln}
\alias{GeneVln}
\title{GeneVln}
\usage{
GeneVln(
  expr = NULL,
  group = NULL,
  title = "x",
  ylab = NULL,
  vlncolor = colorRampPalette(pal_rickandmorty()(12))(12),
  errorbar = FALSE,
  max_y = 3,
  pvalue = TRUE
)
}
\arguments{
\item{expr}{expression matrix including several genes}

\item{group}{group name}

\item{title}{plot title}

\item{ylab}{Y-axis title}

\item{vlncolor}{color of violin}

\item{errorbar}{logical value, whether to add errorbar}

\item{max_y}{Y-axis maximum value.}

\item{pvalue}{logical value, whether to add pvalue label}
}
\description{
violin plot of gene expression in different branches.
}
