% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_blank_coord2}
\alias{get_blank_coord2}
\title{get_blank_coord2}
\usage{
get_blank_coord2(p, x_cut = NULL, y_cut = NULL)
}
\arguments{
\item{p}{object of ggplot}

\item{x_cut}{range of x}

\item{y_cut}{range of x}
}
\description{
Find a blank area within the given interval in the figure and label it pval
}
