% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_pval_df}
\alias{get_pval_df}
\title{get_pval_df}
\usage{
get_pval_df(dff, attr_name, value_name, group_name, test_method, seq = 1)
}
\arguments{
\item{dff}{Original data frame}

\item{attr_name}{column name for splitting the data frame}

\item{value_name}{Numeric column for plotting.}

\item{group_name}{column name for data grouping within each data frame}

\item{test_method}{statistical test method}

\item{seq}{the interval between p-values, used for plotting}
}
\description{
Retrieve the data frame used for adding significance labels (geom_signif).
}
