% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{GetAllParamMeanChain}
\alias{GetAllParamMeanChain}
\title{GetAllParamMeanChain}
\usage{
GetAllParamMeanChain(
  object,
  iter_range = NULL,
  cores = 1,
  aspect = "pse",
  save = F
)
}
\arguments{
\item{object}{mgpfact object}

\item{iter_range}{Iteration range, default is c(1, 3000)}

\item{cores}{threads}

\item{aspect}{pse or track sub object}

\item{save}{whether save to object}
}
\description{
Get the average value of all parameters on each chain
}
