% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{GetBinTree}
\alias{GetBinTree}
\title{GetBinTree}
\usage{
GetBinTree(object, save = TRUE)
}
\arguments{
\item{object}{MGPfact object}

\item{save}{Logical value, whether save the result in object}
}
\description{
Get the deconstructed multiple trajectories,
and display them through a single-fork binary tree
}
