% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{GetGeneCurve}
\alias{GetGeneCurve}
\title{gaussian_process_regressionDF}
\usage{
GetGeneCurve(object, genes = NULL, n.samples = 30)
}
\arguments{
\item{object}{MGPfact object}

\item{genes}{a character contain gene names}

\item{n.samples}{The number of points included in the regression results}
}
\description{
Use Gaussian process regression to calculate the smooth
expression of any gene/trajectory score along a trajectory.
}
