% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{GetIterCor}
\alias{GetIterCor}
\title{GetIterCor}
\usage{
GetIterCor(object, iteration_list = NULL, chains = NULL)
}
\arguments{
\item{object}{mgpfact object}

\item{iteration_list}{a list contain iteration range, eg. list(c(1, 2000),c(1,4000))}

\item{chains}{The Markov chains for parameter statistics;
if not specified, statistics will be conducted for all}
}
\description{
use this function, obtain correlations under different
iterations of different chains
}
