% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pse.R
\name{GetMURPMapLabel}
\alias{GetMURPMapLabel}
\title{GetMURPMapLabel}
\usage{
GetMURPMapLabel(object, labels = NULL)
}
\arguments{
\item{object}{MGPfact object}

\item{labels}{Cell attribute labels to map in MURP}
}
\description{
map the existing attribute labels about cells to murp

Arguments:
}
