% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{GetNPC}
\alias{GetNPC}
\title{RUNDM}
\usage{
GetNPC(pca_result = NULL, sd_cutoff = 1)
}
\arguments{
\item{pca_result}{pca result from prcomp}

\item{sd_cutoff}{Cut-off threshold for principal component variance}

\item{object}{MGPfact object}

\item{neigen}{number of components}

\item{method}{the method of compute dist matrix}

\item{cores}{the number of threads that can be used to calculate distances}
}
\description{
Perform diffusion map dimensionality reduction based on the expression matrix

Get the number of principal components, default is sd > 1
}
