% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{GetParamMeanChain}
\alias{GetParamMeanChain}
\title{GetParamMeanChain}
\usage{
GetParamMeanChain(
  mamba = NULL,
  murp = NULL,
  param = NULL,
  whichPos = "all",
  iter_range = c(1, 3000),
  cores = 1
)
}
\arguments{
\item{mamba}{mamba object}

\item{murp}{murp result}

\item{param}{parameter name}

\item{whichPos}{'all' / c(1,2,3)}

\item{iter_range}{Iteration range, default is c(1, 3000)}

\item{cores}{threads}
}
\description{
Get the average value of each chain for any parameter
}
