% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{getParams,MGPfact-method}
\alias{getParams,MGPfact-method}
\title{getParams}
\usage{
\S4method{getParams}{MGPfact}(x, slotName = NULL)
}
\arguments{
\item{x}{An MGPfact object}

\item{slotName}{slot name}
}
\description{
getParams
}
