% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pse.R
\name{GetPseSdf}
\alias{GetPseSdf}
\title{GetPseSdf}
\usage{
GetPseSdf(
  object,
  param_meanvalue = NULL,
  unified_direction = FALSE,
  rm_adjust_chain = FALSE,
  rev = FALSE
)
}
\arguments{
\item{object}{MGPfact object}

\item{param_meanvalue}{The data frame containing the mean values of each parameter,
obtainable through 'GetAllParamMeanChain'}

\item{unified_direction}{Logical value, indicating whether to
automatically adjust the representation of differentiation direction by pseudotime}

\item{rm_adjust_chain}{Logical value, indicating whether to discard some low-quality chains}

\item{rev}{Whether to perform direction reversal based on pseudotime}
}
\description{
Organize the results obtained from 'RunningmodMGPpseudoT'
}
