% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{GetSigDF}
\alias{GetSigDF}
\title{GetSigDF}
\usage{
GetSigDF(cs, traj, ct_method)
}
\arguments{
\item{cs}{MGPfact-gene object}

\item{traj}{trajectory number

sig = TRUE
ct_method = "lm"
sig_method = "pvalue"
sig_th = 0.05}

\item{ct_method}{dge method}
}
\description{
Obtain a data frame of significant genes identified by different
differential analysis methods, including results for p-value and adjusted p-value.
}
