% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{GetSigGene}
\alias{GetSigGene}
\title{GetSigGene}
\usage{
GetSigGene(cs, traj, ct_method, sig_method, sig_label, sig_th)
}
\arguments{
\item{cs}{MGPfact-gene object}

\item{traj}{trajectory index}

\item{ct_method}{dge method}

\item{sig_method}{statistics label}

\item{sig_label}{dge factor}

\item{sig_th}{filter threshold

sig = TRUE
ct_method = "lm"
sig_method = "pvalue"
sig_th = 0.05}
}
\description{
Filter genes identified as significant by different differential analysis methods
}
