% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{GetSigmaFromJulia}
\alias{GetSigmaFromJulia}
\title{GetSigmaFromJulia}
\usage{
GetSigmaFromJulia(object, load = FALSE, julia_home = NULL, init = T)
}
\arguments{
\item{object}{MGPfact object}

\item{load}{logical value, whether to load julia's environment}

\item{julia_home}{Julia's bin path}

\item{init}{initial values}
}
\description{
get the covariance matrix
}
