% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{GetTbTree}
\alias{GetTbTree}
\title{GetTbTree}
\usage{
GetTbTree(object, save = TRUE)
}
\arguments{
\item{object}{MGPfact object}

\item{save}{Logical value, whether save tbtree result in mgpfact}
}
\description{
get tbtree result of murp
}
