% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{GetTbTreeAllpoint}
\alias{GetTbTreeAllpoint}
\title{GetTbTreeAllpoint}
\usage{
GetTbTreeAllpoint(object, save = TRUE, labels = NULL)
}
\arguments{
\item{object}{MGPfact object}

\item{save}{Logical value, whether save tbtree result in mgpfact}

\item{labels}{some attribute about cells}
}
\description{
add all point in tbtree
}
