% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{GetTbTreeAllResult}
\alias{GetTbTreeAllResult}
\title{GetTbTreeAllResult}
\usage{
GetTbTreeAllResult(object)
}
\arguments{
\item{object}{MGPfact object}
}
\description{
extract consensus tree result, which including all cells
}
