% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{GetTbTreeDrq}
\alias{GetTbTreeDrq}
\title{GetTbTree}
\usage{
GetTbTreeDrq(object, save = TRUE)
}
\arguments{
\item{object}{MGPfact object}

\item{save}{Logical value, whether save the result in object}
}
\description{
Get a multi-forked binary-tree by merging different trajectories
}
