% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{GetTrackSdfFromOptim}
\alias{GetTrackSdfFromOptim}
\title{GetTrackSdf}
\usage{
GetTrackSdfFromOptim(object = NULL, iter_range = NULL, save = TRUE)
}
\arguments{
\item{object}{MGPfact object}

\item{iter_range}{Retain the results of parameter optimization
within a specified iteration range}

\item{save}{logical value, whether to save locally}
}
\description{
Organize the results obtained from 'RunningmodMGPtrack'
}
