% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{GetTrajectoryCurve}
\alias{GetTrajectoryCurve}
\title{GetTrajectoryCurve}
\usage{
GetTrajectoryCurve(object, n.samples = 30)
}
\arguments{
\item{object}{MGPfact object}

\item{n.samples}{The number of points included in the regression results}
}
\description{
Smooth any trajectory score using Gaussian process regression
}
