% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{GetWeightGene}
\alias{GetWeightGene}
\title{GetWeightGene}
\usage{
GetWeightGene(
  object,
  method = c("weight_cut", "top_ratio", "top_number"),
  trajectory
)
}
\arguments{
\item{track}{track object}

\item{q_seq}{}
}
\description{
GetWeightGene
}
