% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ImportPseResult}
\alias{ImportPseResult}
\title{ImportPseSimToR}
\usage{
ImportPseResult(object, init = FALSE, julia_home = FALSE)
}
\arguments{
\item{object}{MGPfact object}

\item{init}{Logical value, whether to import the initial value.
it can be viewed in the directory: 2_pseudotime/2.1_julia_result/inits.jld2}

\item{julia_home}{julia bin path}
}
\description{
Import the pseudotime object optimized by julia into R
}
