% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{ImportTrackResult}
\alias{ImportTrackResult}
\title{ImportOptimR}
\usage{
ImportTrackResult(object, optim_result = NULL, hasinits = TRUE)
}
\arguments{
\item{object}{MGPfact object}

\item{optim_result}{tracking result from optim}

\item{hasinits}{Logical value, whether has initial values}
}
\description{
import trajectory scores in mgpfact object
}
