% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{IntersectSet}
\alias{IntersectSet}
\title{IntersectSet}
\usage{
IntersectSet(set_list = NULL, num = 3)
}
\arguments{
\item{set_list}{List containing multiple sets.}

\item{num}{number of sets.}
}
\description{
Intersection among any num sets selected from multiple sets
}
