% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{LogNormalize}
\alias{LogNormalize}
\title{LogNormalize}
\usage{
LogNormalize(data, scale.factor = 10000, margin = 2L, verbose = TRUE, ...)
}
\arguments{
\item{data}{count matrix}

\item{scale.factor}{scale factor}

\item{margin}{Which dimension to standardize by}

\item{verose}{logical value, whether to output progress}
}
\description{
normalize data
}
