% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{MapMURPLabelToAll}
\alias{MapMURPLabelToAll}
\title{MapMURPLabelToAll}
\usage{
MapMURPLabelToAll(vecc, orig)
}
\arguments{
\item{vecc}{new label}

\item{orig}{murp label for each cell}
}
\description{
mapping MURP label to all cells
}
