% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{MGPfact-class}
\alias{MGPfact-class}
\alias{MGPfact}
\title{The MGPfact-class}
\description{
The MGPfact-class
}
\section{Slots}{

\describe{
\item{\code{assay}}{expression matrix}

\item{\code{OptimResult}}{julia results}

\item{\code{MURP}}{murp results}

\item{\code{DimReducs}}{dimensionality reduction results}

\item{\code{MetaData}}{meta data}

\item{\code{Date}}{object creation date}

\item{\code{BeautGene}}{results related to genes}

\item{\code{Tree}}{trajectory tree}

\item{\code{GPR}}{gaussian process results}

\item{\code{Command}}{execution log}
}}

