% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{MURPPCA}
\alias{MURPPCA}
\title{MURPPCA}
\usage{
MURPPCA(object, pca.center = FALSE, pca.scale = FALSE)
}
\arguments{
\item{object}{MGPfact object}

\item{pca.center}{MURPPCA parameter: a logical value indicating whether the variables should be
shifted to be zero centered}

\item{pca.scale}{MURPPCA parameter: a logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place}
}
\description{
Perform Principal Component Analysis on the results of MURP downsampling
}
