% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{ObjectFunction}
\alias{ObjectFunction}
\title{ObjectFunction}
\usage{
ObjectFunction(par, Yx, S_list, debug_file)
}
\arguments{
\item{par}{Initial values for the parameters to be estimated}

\item{Yx}{Matrix required for calculating trajectory scores}

\item{S_list}{a list, covariance matrix for all trajectories}

\item{debug_file}{log files}
}
\description{
The objective function for calculating trajectory scores
}
