% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_plot.R
\name{PCANestedGridPlot}
\alias{PCANestedGridPlot}
\title{PCANestedGridPlot}
\usage{
PCANestedGridPlot(pca_result = NULL, sd_cutoff = 1, max_pc = NULL)
}
\arguments{
\item{pca_result}{pca result}

\item{sd_cutoff}{choose the right number of pc according sd}

\item{max_pc}{maximum number of pc}
}
\description{
Partial zoom
}
