% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{plot_pie}
\alias{plot_pie}
\title{pie plot for every row of data.frame}
\usage{
plot_pie(i, tab, levs, size = 0.2, color = NA, col_values = NULL)
}
\arguments{
\item{i}{Extract the attributes of the i-th column}

\item{tab}{DataFrame related to the property}

\item{levs}{facter levels}

\item{size}{point size}

\item{color}{point outline color}

\item{col_values}{point fill color}
}
\description{
Pie chart composition
}
