% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotBinBif}
\alias{PlotBinBif}
\title{PlotBinBif}
\usage{
PlotBinBif(object, plot_label = FALSE)
}
\arguments{
\item{object}{MGPfact object}

\item{plot_label}{logical value, whether to plot label}
}
\description{
Visualize bifurcation on the binary-tree
}
