% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotBinLabel}
\alias{PlotBinLabel}
\title{PlotBinLabel}
\usage{
PlotBinLabel(object, plot_label = FALSE, labels = NULL)
}
\arguments{
\item{object}{MGPfact object}

\item{plot_label}{logical value, whether to plot label}

\item{labels}{cell property labels}
}
\description{
Visualize cell property labels on the binary-tree
}
