% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_plot.R
\name{PlotLabelMerge}
\alias{PlotLabelMerge}
\title{PlotLabelMerge}
\usage{
PlotLabelMerge(object = NULL, labels = NULL, width = 17)
}
\arguments{
\item{object}{MGPfact object}

\item{labels}{attributes in metdata}

\item{width}{canvas width}
}
\description{
plot label in tSNE/UMAP/DM/PCA merge
}
