% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_plot.R
\name{PlotLabelSplit}
\alias{PlotLabelSplit}
\title{PlotLabelSplit}
\usage{
PlotLabelSplit(object = NULL, labels = NULL, width = 9)
}
\arguments{
\item{object}{MGPfact object}

\item{labels}{attributes in metdata}

\item{width}{canvas width}
}
\description{
plot label in tSNE/UMAP/DM/PCA split
}
