% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{PlotParamLogpdf}
\alias{PlotParamLogpdf}
\title{PlotParamLogpdf}
\usage{
PlotParamLogpdf(object = NULL, iter_range = NULL)
}
\arguments{
\item{object}{mgpfact object}

\item{iter_range}{Iteration range, default is c(1, 3000)}
}
\description{
Get the average value of all parameters on each chain
}
