% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{PlotParamSampling}
\alias{PlotParamSampling}
\title{PlotParamSampling}
\usage{
PlotParamSampling(object, iter_range = NULL, params = "T")
}
\arguments{
\item{object}{mgpfact object}

\item{iter_range}{Iteration range, default is c(1, 3000)}

\item{params}{parameter name, can be a character vector.}
}
\description{
Get the average value of all parameters on each chain
}
