% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pse.R
\name{PlotPCPseBif}
\alias{PlotPCPseBif}
\title{PlotPCPseBif}
\usage{
PlotPCPseBif(object, lm = FALSE)
}
\arguments{
\item{object}{MGPfact object}

\item{lm}{Logical value, indicating whether to add a linear regression curve}
}
\description{
Visualize bifurcation effects on PC and pseudotime
}
