% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pse.R
\name{PlotPCPseLabel}
\alias{PlotPCPseLabel}
\title{PlotPCPseLabel}
\usage{
PlotPCPseLabel(object, labels, lm = FALSE)
}
\arguments{
\item{object}{MGPfact object}

\item{labels}{attributes need to plot}

\item{lm}{Logical value, indicating whether to add a linear regression curve}
}
\description{
Visualize different labels on PC and pseudotime with
}
