% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotPieBinLabel}
\alias{PlotPieBinLabel}
\title{PlotPieBinLabel}
\usage{
PlotPieBinLabel(object, labels = NULL, size = 0.01, color = "black")
}
\arguments{
\item{object}{MGPfact object}

\item{labels}{cell property labels}

\item{size}{point size}

\item{color}{Point outline color}
}
\description{
Visualize cell property labels on the binary-tree,
using the proportion of cell types in the MURP subset to represent the points.
}
