% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotPieConsensusAllLabel}
\alias{PlotPieConsensusAllLabel}
\title{PlotPieTbAllLabel}
\usage{
PlotPieConsensusAllLabel(
  object,
  labels = NULL,
  size = 0.2,
  color = "black",
  sub_scale_factor = 0.7,
  cell_size = 1,
  cell_alpha = 0.5
)
}
\arguments{
\item{object}{MGPfact object}

\item{labels}{cell property labels}

\item{size}{pie size}

\item{color}{Point outline color}

\item{sub_scale_factor}{Sub view's size scaling factor}

\item{cell_size}{point size}

\item{cell_alpha}{point alpha}
}
\description{
Visualize cell property labels on the main path of the consensus trajectory,
using the proportion of cell types in the MURP subset to represent the points.
}
