% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotPieTbLabel}
\alias{PlotPieTbLabel}
\title{PlotPieTbLabel}
\usage{
PlotPieTbLabel(object, labels = NULL, size = 0.2, color = "black")
}
\arguments{
\item{object}{MGPfact object}

\item{labels}{cell property labels}

\item{size}{point size}

\item{color}{Point outline color}
}
\description{
Visualize cell property labels on the main path of the tb-tree,
using the proportion of cell types in the MURP subset to represent the points.
}
