% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotTbBif}
\alias{PlotTbBif}
\title{PlotTbBif}
\usage{
PlotTbBif(object, plot_label = FALSE)
}
\arguments{
\item{object}{MGPfact object}

\item{plot_label}{logical value, whether to plot label}
}
\description{
Visualize the branching of independent differentiation trajectories on the tb-tree
}
