% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotTbLabel}
\alias{PlotTbLabel}
\title{PlotTbLabel}
\usage{
PlotTbLabel(object, labels = NULL)
}
\arguments{
\item{object}{MGPfact object}

\item{labels}{cell property labels}
}
\description{
Visualize cell property labels on the tb-tree
}
