% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotTbMainLabel}
\alias{PlotTbMainLabel}
\title{PlotTbMainLabel}
\usage{
PlotTbMainLabel(object, labels = NULL)
}
\arguments{
\item{object}{MGPfact object}

\item{labels}{cell property labels}
}
\description{
Visualize cell property labels on the main path of the consensus trajectory
}
