% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotTbMainPse}
\alias{PlotTbMainPse}
\title{PlotTbMainPse}
\usage{
PlotTbMainPse(object, col = "blue")
}
\arguments{
\item{object}{MGPfact object}

\item{col}{color}
}
\description{
plot pseudotime on the main path of the consensus trajectory
}
