% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotTbPse}
\alias{PlotTbPse}
\title{PlotTbPse}
\usage{
PlotTbPse(object, plot_label = FALSE)
}
\arguments{
\item{object}{MGPfact object}

\item{plot_label}{logical value, whether to plot label}
}
\description{
Visualize pseudotime on the binary-tree
}
