% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{PseOptimResult-class}
\alias{PseOptimResult-class}
\alias{PseOptimResult}
\title{The PseOptimResult-class}
\description{
The PseOptimResult-class
}
\section{Slots}{

\describe{
\item{\code{sample_value}}{sampling results for each iteration}

\item{\code{dimnames}}{names of sampling value}

\item{\code{burnin}}{burning period}

\item{\code{chains}}{chain number}

\item{\code{iter}}{iteration number}

\item{\code{hasinits}}{is there an initial value}

\item{\code{inits}}{initial value}

\item{\code{iter_chain_cor}}{correlation between pseudotime and
sampling time points across different iterations and Markov chains.}

\item{\code{t_pred}}{pseudotime of murp}

\item{\code{sortindex_cor}}{(ready to delete)}

\item{\code{diagnostic}}{convergence judgment results}

\item{\code{logpdf}}{The objective function value of each optimization}
}}

