% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pval_lab}
\alias{pval_lab}
\title{pval_lab}
\usage{
pval_lab(pval, lab = T, hide_ns = F)
}
\arguments{
\item{pval}{P value}

\item{lab}{Whether to add significance marks *}

\item{hide_ns}{Whether to hide labels that are not significant}
}
\description{
Transforming p-values for visualization.
}
