% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ReadPseSim}
\alias{ReadPseSim}
\title{ReadPseSim}
\usage{
ReadPseSim(sim = NULL, julia_home = NULL, init = FALSE)
}
\arguments{
\item{sim}{julia result}

\item{julia_home}{Julia's bin path}

\item{init}{Logical value, whether to import the initial value.
it can be viewed in the directory: 2_pseudotime/2.1_julia_result/inits.jld2}
}
\description{
read and convert the object optimized by julia into R
}
