% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RenameCluster}
\alias{RenameCluster}
\title{RenameCluster}
\usage{
RenameCluster(vector = NULL, replace = NULL)
}
\arguments{
\item{vector}{a vector to replace}

\item{replace}{the content to be replaced}
}
\description{
A replacement tool, such as replacing all "a" in the vector vector with "1"
RenameCluster(vector = c(1,1,1,2), replace = c("1"="a","2"="b"))
}
