% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{rotate_point}
\alias{rotate_point}
\title{rotate_point}
\usage{
rotate_point(x, y, angle)
}
\arguments{
\item{x}{X-axis}

\item{y}{Y-axis}

\item{angle}{Rotation angle}
}
\description{
Rotate coordinates counterclockwise.
}
