% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{RunningmodMGPpseudoT}
\alias{RunningmodMGPpseudoT}
\title{RunningmodMGPpseudoT}
\usage{
RunningmodMGPpseudoT(object, julia_home, seed = 723, cores = 1)
}
\arguments{
\item{object}{MGPfact object}

\item{julia_home}{Julia's bin path}

\item{seed}{random seed, default is 723}

\item{cores}{number of threads}
}
\description{
trajectory parameters estimation through MCMC sampling
}
