% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{RunningmodMGPtrack}
\alias{RunningmodMGPtrack}
\title{RunningmodMGPtrack}
\usage{
RunningmodMGPtrack(object, iter = 100, cores = 1, seed = 723)
}
\arguments{
\item{object}{MGPfact object}

\item{iter}{optim iterations}

\item{cores}{number of threads}

\item{seed}{random seed}
}
\description{
calculating trajectory scores
}
