% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RunTime}
\alias{RunTime}
\title{RunTime}
\usage{
RunTime(t1, t2, timeunits = "secs_mins")
}
\arguments{
\item{t1}{start}

\item{t2}{end}

\item{timeunits}{time units}
}
\description{
Running time of statistical algorithm
}
