% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{RUNUMAP}
\alias{RUNUMAP}
\title{RUNUMAP}
\usage{
RUNUMAP(object, npc = 1:30)
}
\arguments{
\item{object}{MGPfact object}

\item{npc}{which pc}
}
\description{
Perform UMAP dimensionality reduction based on the principal component analysis
}
