% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{Save2CS}
\alias{Save2CS}
\title{Save2CS}
\usage{
Save2CS(obj, base, name, result)
}
\arguments{
\item{obj}{MGPfact-gene object}

\item{base}{storage location}

\item{name}{name of storage location}

\item{result}{Results to be stored}
}
\description{
save result to mgpfact-gene object
}
