% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{SaveMURPDatToJulia}
\alias{SaveMURPDatToJulia}
\title{SaveMURPDatToJulia}
\usage{
SaveMURPDatToJulia(object, murp_pc_number = 3)
}
\arguments{
\item{object}{MGPfact object}

\item{murp_pc_number}{the number of principal components of
the expression matrix after MURP downsampling}
}
\description{
extract murp data to julia environment
}
