% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SetSettings}
\alias{SetSettings}
\title{set parameters of settings}
\usage{
SetSettings(
  object,
  murp_pc_number = 3,
  trajectory_number = 3,
  pse_optim_iterations = 10,
  start_murp = 1,
  chains_number = 10
)
}
\arguments{
\item{object}{MGPfact object}

\item{murp_pc_number}{the number of principal components of the expression matrix after MURP downsampling}

\item{trajectory_number}{the number of trajectories to be deconstructed}

\item{pse_optim_iterations}{the number of optimization iterations to predict pseudotime}

\item{start_murp}{root point, If not required, set it to a number larger than the number of murps}

\item{chains_number}{the number of Markov chains}
}
\description{
Set parameters necessary in optimization
}
