% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{subString}
\alias{subString}
\title{subString}
\usage{
subString(strings, idx, sep = NA)
}
\arguments{
\item{strings}{character}

\item{idx}{extract the idx-th}

\item{sep}{delimiter}
}
\description{
split string with sep, get idx substring
}
