% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{TrackOptimResult-class}
\alias{TrackOptimResult-class}
\alias{TrackOptimResult}
\title{The TrackOptimResult-class}
\description{
The TrackOptimResult-class
}
\section{Slots}{

\describe{
\item{\code{sample_value}}{sampling results for each iteration}

\item{\code{dimnames}}{names of sampling value}

\item{\code{burnin}}{burning period}

\item{\code{chains}}{chain number}

\item{\code{iter}}{iteration number}

\item{\code{hasinits}}{is there an initial value}

\item{\code{inits}}{inits}
}}

