% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{WeightGeneFilter}
\alias{WeightGeneFilter}
\title{WeightGeneFilter}
\usage{
WeightGeneFilter(
  object,
  iter_range = NULL,
  method = c("top_ratio", "quantile", "confidence", "top_number", "weight_cut"),
  conf.int = 0.95,
  q_seq = 0.25,
  top_ratio = 0.1,
  top_number = 500,
  weight_cut = 0.03
)
}
\arguments{
\item{object}{MGPfact object}

\item{iter_range}{Retain the results of parameter optimization
within a specified iteration range}

\item{method}{filtering methods
c("top_ratio","quantile","confidence","top_number","weight_cut")}

\item{conf.int}{patameters for the confidence: select genes with the highest weights based on confidence}

\item{q_seq}{parameters for the quantile method: quantile intervals}

\item{top_ratio}{parameters for the top_ratio: select genes with the highest weights based on proportion}

\item{top_number}{patameters for the top_number: select genes with the highest weights based on quantity}

\item{weight_cut}{patameters for the weight_cut: select genes with the highest weights based on weight}
}
\description{
Filter gene weights
}
