% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{WriteGene2TXT}
\alias{WriteGene2TXT}
\title{SaveSigSDF}
\usage{
WriteGene2TXT(cs, ct_method, sig_method, sig_label, sig_th)
}
\arguments{
\item{cs}{MGPfact-gene object}

\item{ct_method}{dge method}

\item{sig_method}{statistics label}

\item{sig_label}{dge factor}

\item{sig_th}{filter threshold
sig = TRUE
ct_method = "lm"
sig_method = "pvalue"
sig_th = 0.05}

\item{ct}{MGPfact object}
}
\description{


Write genes with significant differences to a txt file
}
