% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{WriteWeightGene}
\alias{WriteWeightGene}
\title{WriteWeightGene}
\usage{
WriteWeightGene(object, method = c("weight_cut", "top_ratio", "top_number"))
}
\arguments{
\item{object}{MGPfact object}

\item{method}{filtering methods
c("top_ratio","quantile","confidence","top_number","weight_cut")}
}
\description{
WriteWeightGene
}
