% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{PlotPieConsensusAllLabel}
\alias{PlotPieConsensusAllLabel}
\title{PlotPieTbAllLabel}
\usage{
PlotPieConsensusAllLabel(
  object,
  labels = NULL,
  size = 0.2,
  color = "black",
  scale_factor = 0.5
)
}
\arguments{
\item{object}{MGPfact object}

\item{labels}{cell property labels}

\item{size}{point size}

\item{color}{Point outline color}
}
\description{
Visualize cell property labels on the main path of the consensus trajectory,
using the proportion of cell types in the MURP subset to represent the points.
}
