% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{RUNtSNE}
\alias{RUNtSNE}
\title{RUNtSNE}
\usage{
RUNtSNE(object, npcs = 1:30)
}
\arguments{
\item{object}{MGPfact object}

\item{npcs}{which pc}
}
\description{
Perform tSNE dimensionality reduction based on the principal component analysis
}
