% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{TrajPlot}
\alias{TrajPlot}
\title{TrajPlot}
\usage{
TrajPlot(
  track_sdf = NULL,
  pointColorValue = NULL,
  indexL = 1,
  title = T,
  g_title = NULL,
  col = NULL,
  c = NULL,
  w = NULL,
  tb = TRUE,
  tb_pred = NULL,
  pointSize = 4,
  pointAlpha = 0.4,
  pointLabel = FALSE,
  pointLabelsize = 3,
  rug = TRUE,
  rugAlpha = 0.3,
  legend = FALSE,
  legend_title = "Bif",
  lm = TRUE,
  lineMethod = loess,
  lineType = "solid",
  lineSize = 1,
  lineAlpha = 0.3,
  se = F,
  span = 0.75,
  formula = y ~ x,
  family = NULL,
  vlineType = "dashed",
  ...
)
}
\arguments{
\item{indexL}{trajectory index}

\item{title}{plot title}

\item{col}{the colors of different branches}

\item{c}{branching column}

\item{w}{trajectory score column}

\item{tb}{logical value, whether label bifurcation line}

\item{tb_pred}{bifurcation point}

\item{pointSize}{point size}

\item{pointAlpha}{point alpha}

\item{pointLabel}{logical value, whether label point}

\item{pointLabelsize}{label size}

\item{rug}{logical value, whether to plot rug}

\item{rugAlpha}{rug alpha}

\item{legend}{locical value, whether to add legend}

\item{legend_title}{legend title}

\item{lm}{whether add lm line}

\item{lineMethod}{linear model methods}

\item{lineType}{line type}

\item{lineSize}{line size}

\item{lineAlpha}{line alpha}

\item{track_df}{dataframe for plot smooth trajectory}
}
\description{
smooth trajectory visualization
}
